requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0,<2.16.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
